Implement dead reckoning algorithms
All accepted dead reckoning algorithms are present inthe switch. However, they are not all fully implemented. The dead reckoning calculations seem to be correct but more testing needs to be performed. Additionally, the values from the dead reckoning other parameters are being read and somewhat used. The quaternion inputs need to be utilized to finish the dead reckoning algorithms.